+++ src/main/java/ddingdong/ddingdongBE/domain/activityreport/entity/Participant.java
@@ -1,4 +1,4 @@
-package ddingdong.ddingdongBE.domain.activityreport.domain;
+package ddingdong.ddingdongBE.domain.activityreport.entity;
 
 import jakarta.persistence.Embeddable;
