+++ src/main/java/ddingdong/ddingdongBE/domain/activityreport/service/ActivityReportService.java
@@ -1,6 +1,6 @@
 package ddingdong.ddingdongBE.domain.activityreport.service;
 
-import ddingdong.ddingdongBE.domain.activityreport.domain.ActivityReport;
+import ddingdong.ddingdongBE.domain.activityreport.entity.ActivityReport;
 import ddingdong.ddingdongBE.domain.club.entity.Club;
 import java.util.List;
