+++ src/main/java/ddingdong/ddingdongBE/domain/activityreport/service/ActivityReportTermInfoServiceImpl.java
@@ -1,6 +1,6 @@
 package ddingdong.ddingdongBE.domain.activityreport.service;
 
-import ddingdong.ddingdongBE.domain.activityreport.domain.ActivityReportTermInfo;
+import ddingdong.ddingdongBE.domain.activityreport.entity.ActivityReportTermInfo;
 import ddingdong.ddingdongBE.domain.activityreport.repository.ActivityReportTermInfoRepository;
 import java.time.LocalDate;
 import java.time.LocalDateTime;
